# openapi.yaml
openapi: 3.0.0
paths:
...
  /auth/login:
    post:
      description: Submit username and password to receive a verification code via email
      operationId: AuthController_login
      parameters: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginDto"
      responses:
        "200":
          description: Verification code sent to email
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: Verification code sent to your email
        "400":
          description: Invalid input data
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponseDto"
        "401":
          description: Invalid credentials
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponseDto"
      summary: Initiate login
      tags: &a5
        - auth
  /auth/verify:
    post:
      description: Submit email and verification code to complete login and receive session cookie
      operationId: AuthController_verify
      parameters: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/VerifyCodeDto"
      responses:
        "200":
          description: Login successful, session cookie set
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: Login successful
                  user:
                    type: object
                    properties:
                      id:
                        type: string
                      username:
                        type: string
                      email:
                        type: string
                      type:
                        type: string
        "400":
          description: Invalid input data
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponseDto"
        "401":
          description: Invalid or expired verification code
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponseDto"
      summary: Verify code and complete login
      tags: *a5
  /auth/logout:
    post:
      description: Destroy the current session
      operationId: AuthController_logout
      parameters: []
      responses:
        "200":
          description: Logged out successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: Logged out successfully
        "401":
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponseDto"
      security:
        - session-id: []
      summary: Logout
      tags: *a5
  /auth/me:
    get:
      description: Get information about the currently authenticated user
      operationId: AuthController_getCurrentUser
      parameters: []
      responses:
        "200":
          description: Current user information
          content:
            application/json:
              schema:
                type: object
                properties:
                  id:
                    type: string
                  username:
                    type: string
                  email:
                    type: string
                  type:
                    type: string
        "401":
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponseDto"
      security:
        - session-id: []
      summary: Get current user
      tags: *a5
info:
  title: Nest API Demo
  description: A comprehensive REST API built with NestJS and MySQL. This API provides endpoints for managing users, employees, and departments with role-based access control.
  version: "1.0"
  contact: {}
...
servers: []
components:
  securitySchemes:
    cookie:
      type: apiKey
      in: cookie
      name: session-id
  schemas:
...
    ErrorResponseDto:
      type: object
      properties:
        statusCode:
          type: number
          description: HTTP status code of the error response
          example: 401
        timestamp:
          type: string
          description: ISO 8601 timestamp of when the error occurred
          example: 2024-01-15T10:30:00.000Z
          format: date-time
        path:
          type: string
          description: Request path that produced the error
          example: /api/employees/123
        response:
          description: Error response payload (string message or an object returned by Nest exceptions)
          oneOf:
            - type: string
              example: Unauthorized
            - type: object
              additionalProperties: true
              example:
                message: Unauthorized
                statusCode: 401
      required:
        - statusCode
        - timestamp
        - path
        - response
...
    LoginDto:
      type: object
      properties:
        username:
          type: string
          description: The username of the user
          example: johndoe
        password:
          type: string
          description: The password of the user
          example: SecurePassword123!
      required:
        - username
        - password
    VerifyCodeDto:
      type: object
      properties:
        email:
          type: string
          description: The email address of the user
          example: john.doe@example.com
          format: email
        code:
          type: string
          description: The verification code sent to the email
          example: "123456"
          minLength: 4
          maxLength: 8
      required:
        - email
        - code